# docs/api/openapi.yaml
{
  "openapi": "3.1.1",
  "info": {
    "title": "StudentOS API",
    "description": "Personal academic operating system — REST API for web and mobile clients.\n\nAuthentication uses HS256 JWT access tokens (short-lived) and refresh tokens stored in the database. Attach the access token as `Authorization: Bearer <token>` or via the `x-access-token` header.\n\nAll authenticated endpoints return the standard envelope from the shared contract: `{ success: true, data: T, meta?: ... }` on success and `{ success: false, error: { code, message } }` on failure.\n\n**Ownership model:** Every resource (academic year, semester, course, task, note, study session, goal, etc.) belongs to a single user. Users can only see and modify their own resources. Unauthorized access returns `FORBIDDEN`. Missing resources return `NOT_FOUND`.",
    "version": "0.1.0",
    "contact": {
      "name": "StudentOS",
      "url": "https://github.com/elyasdemamu/StudentOS"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Base path for all endpoints"
    }
  ],
...
  "paths": {
...
    "/courses": {
      "get": {
        "tags": ["Courses"],
        "summary": "List courses",
        "description": "Returns the authenticated user's courses. Supports filtering by `academicYearId`, `semesterId`, `status`, text search on `name`/`code`, and pagination. Returns the total count in `meta`.",
        "operationId": "listCourses",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "academicYearId",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "in": "query",
            "name": "semesterId",
            "required": false,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": { "$ref": "#/components/schemas/CourseStatus" }
          },
          {
            "in": "query",
            "name": "search",
            "description": "Search by name or code.",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list with total count",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse_CourseList" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Courses"],
        "summary": "Create a course",
        "description": "Creates a new course for the authenticated user. Optionally associate with an academic year and/or semester. The `code` field is optional (e.g. \"CS-301\").",
        "operationId": "createCourse",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCourseRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse_Course" }
              }
            }
          }
        }
      }
    },
    "/courses/{id}": {
      "get": {
        "tags": ["Courses"],
        "summary": "Get a course",
        "description": "Returns a single course by ID. Only the owning user can access it.",
        "operationId": "getCourse",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Course",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse_Course" }
              }
            }
          },
          "404": {
            "description": "Course not found or not owned by user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Courses"],
        "summary": "Update a course",
        "description": "Partially updates a course. Only fields provided are changed. Only the owning user can update.",
        "operationId": "updateCourse",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCourseRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse_Course" }
              }
            }
          },
          "404": { "description": "Not found or not owned" }
        }
      },
      "delete": {
        "tags": ["Courses"],
        "summary": "Delete a course",
        "description": "Soft-deletes a course (sets `deletedAt`). Only the owning user can delete. Associated tasks are NOT cascaded — they remain but orphaned (the owning user retains them).",
        "operationId": "deleteCourse",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse_Boolean" }
              }
            }
          },
          "404": { "description": "Not found or not owned" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 JWT access token. Send as `Authorization: Bearer <token>` or `x-access-token: <token>`."
      }
    },
    "schemas": {
...
      "CreateCourseRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 200, "example": "Database Systems" },
          "code": { "type": "string", "nullable": true, "maxLength": 50, "example": "CS-301" },
          "academicYearId": { "type": "string", "format": "uuid", "nullable": true },
          "semesterId": { "type": "string", "format": "uuid", "nullable": true },
          "instructor": { "type": "string", "nullable": true, "maxLength": 200 },
          "description": { "type": "string", "nullable": true, "maxLength": 2000 },
          "credits": { "type": "number", "minimum": 0, "maximum": 100, "default": 0 },
          "color": { "type": "string", "nullable": true, "description": "Hex color for the course (e.g. `#3B82F6`). Used in calendar and dashboard." },
          "status": { "$ref": "#/components/schemas/CourseStatus" }
        },
        "additionalProperties": false
      },
      "UpdateCourseRequest": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 200 },
          "code": { "type": "string", "nullable": true },
          "academicYearId": { "type": "string", "format": "uuid", "nullable": true },
          "semesterId": { "type": "string", "format": "uuid", "nullable": true },
          "instructor": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "credits": { "type": "number", "minimum": 0, "maximum": 100 },
          "color": { "type": "string", "nullable": true },
          "status": { "$ref": "#/components/schemas/CourseStatus" }
        },
        "additionalProperties": false
      },
      "CourseStatus": {
        "type": "string",
        "enum": ["UPCOMING", "ACTIVE", "COMPLETED"],
        "default": "ACTIVE",
        "example": "ACTIVE"
      },
...
    }
  }
}